Updated processed data to accomodate for recording errors
Used the Mode helper function as previously used in player_means to set players' reported year and other categorical variables to the mode of whatever they said. This applies to all processed data but not unprocessed data. Also removed the first period in file paths in order to make them work "../whatever" became "./whatever"–which is weird because they worked fine before