Better piecewise regression
Implemented a better piecewise regression that forces the result to be continuous, because otherwise the probability calculations were going to be super funky. As an added benefit, the graphs are way nicer looking now. Also added a way to graph only the best cutoff, and a helper function to predict the probability from a given cutoff