Ran 13.5 hour program
Ran the 13.5 hour program, using max of training and strategy sessions for mutables, and got answers for best teams. Saved to CSV so we don't have to run this again. Notably, the two best teams (for undefeated chance and average wins) share all but two players. Also, might do some more precise runs on the top 10 of each to guarantee that we have the correct order.